// src/app/shared/models/call-user.ts
export class CallUser{
    id: string;
    call_id: string;
    user_id: string;
    required: string;
    accept_status: string;
    date_modified: Date;
    deleted: boolean;
}

export class CallUserJSon{
    id: string;
    callId: string;
    userId: string;
    required: string;
    acceptStatus: string;
    dateModified: Date;
    deleted: boolean;
}